PYTHON-2744 Run LB tests against non-LB clusters
