+++ test/load_balancer/test_load_balancer.py
@@ -14,38 +14,12 @@
 
 """Test the Load Balancer unified spec tests."""
 
-import os
 import sys
 
 sys.path[0:0] = [""]
 
-from test import unittest, IntegrationTest, client_context
-from test.utils import get_pool
-from test.unified_format import generate_test_classes
-
-# Location of JSON test specifications.
-TEST_PATH = os.path.join(
-    os.path.dirname(os.path.realpath(__file__)), 'unified')
-
-# Generate unified tests.
-globals().update(generate_test_classes(TEST_PATH, module=__name__))
-
-
-class TestLB(IntegrationTest):
-    @client_context.require_load_balancer
-    def test_unpin_committed_transaction(self):
-        pool = get_pool(self.client)
-        with self.client.start_session() as session:
-            with session.start_transaction():
-                self.assertEqual(pool.active_sockets, 0)
-                self.db.test.insert_one({}, session=session)
-                self.assertEqual(pool.active_sockets, 1)  # Pinned.
-            self.assertEqual(pool.active_sockets, 1)  # Still pinned.
-        self.assertEqual(pool.active_sockets, 0)  # Unpinned.
-
-    def test_client_can_be_reopened(self):
-        self.client.close()
-        self.db.test.find_one({})
+from test import unittest
+from test.test_load_balancer import *
 
 
 if __name__ == "__main__":
