+++ test/load_balancer/test_versioned_api.py
@@ -13,10 +13,10 @@
 # limitations under the License.
 
 import sys
-import unittest
 
 sys.path[0:0] = [""]
 
+from test import unittest
 from test.test_versioned_api import *
 
 if __name__ == '__main__':
